renamed typeHelper.ts to types.ts
